Rename http/http.ts to http/server.ts
Remove http/mod.ts
