+++ http/file_server.ts
@@ -10,7 +10,7 @@ import {
   ServerRequest,
   setContentLength,
   Response
-} from "./mod.ts";
+} from "./server.ts";
 import { cwd, DenoError, ErrorKind, args, stat, readDir, open } from "deno";
 import { extname } from "../fs/path.ts";
 import { contentType } from "../media_types/mod.ts";
