--- http/mod.ts
@@ -1,7 +1,0 @@
-export {
-  serve,
-  listenAndServe,
-  Response,
-  setContentLength,
-  ServerRequest
-} from "./http.ts";
